Hook up some control flow world script ops.

+++ src/world_script/world_action_funcs.rs
@@ -1,4 +1,4 @@
-pub fn action_func_as_string(address: u16) -> String {
+pub fn action_func_as_string(address: u64) -> String {
     match address {
         0x0F38 => String::from("CODE_FN_C20F38"),
         0x0F63 => String::from("fn_overworld_execute_next_event_commands"),
